Move management of starting and stopping stratisd out of the test class
It doesn't need to have access to the unittest class's self to work.
Combine these actions in a _Service class.

Make separate methods for stopping the daemon and removing the devices.
Generalize _process_exists so that it returns all processes that match the
name.  Use psutil to clean up in the cleanup method.

Signed-off-by: mulhern <amulhern@redhat.com>