Use a context manager to handle starting and stopping the daemon
It helps make it clearer what the test is doing.
Drop the some_existing parameter from _test_driver. If this parameter
is set to True, it causes the daemon to be taken up and down while
devices are being hotplugged during the step when all but the last
device is added to the pool. This must be intended to test how the engine
deals with assembling a whole bunch of incomplete pools at once.
We can test this instead, just once, when all but the last device have
been added.

Signed-off-by: mulhern <amulhern@redhat.com>