Change log entry level and message in identify_block_device
This method is now only invoked when an event is caught or by a test-only
method. An event on an already taken care of device can be caught a lot
without significance to stratisd, so this is not worth logging at the info
level. There is another location in the code where the event of an actual
action being taken because the event of a Stratis device is found is
logged, so there is information enough in the logs.

Signed-off-by: mulhern <amulhern@redhat.com>