Do not dump lots and lots of text on certain failures in tests
It is way too much to read through in logs.

Move _stratis_id.py and code that only it depends on to the stratisd
scripts directory.

_stratis_id.py was originally written for two reasons.
* It allowed dumping out lots of signature block information if udev tests
failed. This was written to attempt to find out the cause of an
intermittent failure in the udev tests. But the output is voluminous and
turned out not to be very useful as it was found that the cause of the
intermittent failures was that occasionally the CreatePool command would
fail, because it was called before the loopbacked devices were properly
set up, but that the test would ignore that failure and proceed. This
problem has been fixed.
* The version of libblkid that recognizes Stratis devices was not yet
generally available, so the _stratis_id.py script provided a programmatic
way to identify Stratis block devices. In most cases, however, blkid
is now just as good.

The reason to keep the stratis_id.py script around in the scripts
directory, is that it provides a structured view of a Stratis signature
buffer that it has read. This is not yet provided by other tools, although
the plan is to ultimately provide it by means of a Rust program that uses
the metadata module which is currently part of stratisd, but which could
be exported as a separate crate. In the meantime, it seems reasonable to
leave the script where it is accessible to us to possibly be helpful in
looking for Stratis metadata on devices. This is not a script that we
would choose to release. Add the usual infrastructure for the newly
introduced directory, but do not try to improve the script at this time.
We can wait until it either gets more use or withers away entirely.

Signed-off-by: mulhern <amulhern@redhat.com>