Swap the order of the action string and the other arguments
Treat the action string as the primary message, and the additional
information provided by the macro arguments as supplementary info.

This should make the message structure pretty readable w/out too
much additional tweaking.

Make the Display implementation messages more self-sufficient.
Generally, this also brings the others better into line with the
one for DestroyAction and work better with the rearrangement.

Emphasize that the name specified in the rename is the new name.

Signed-off-by: mulhern <amulhern@redhat.com>