Remove process_unowned_device in identify module
This was previously invoked if udev identified a block device as unowned
in identify block devices.

The new version of stratisd now required a version of libblkid that
can identify Stratis devices.

Expect that udev will be succesful in identifying Stratis devices, by
setting ID_FS_TYPE when they appear, and drop the additional checks for
unowned devices.

Signed-off-by: mulhern <amulhern@redhat.com>