Rework disk scanning strategy in tests
This commit switches from reading the disk contents into a buffer to
using mmap to avoid large allocations. Both loopback device tests have
been removed due to inconsistencies found between the backing file and the
loopback block device. Testing whether a device is encrypted or not
using loopback devices will therefore always return true even if they
are not encrypted because the loopback block device will not return the
actual data in the file that has been written to.