Rework crypt API to be more granular
This commit changes the API so that it distinguishes between
setup/can_setup (checking that the metadata is well-formed Stratis
metadata for encrypted devices) and can_unlock (checking that the
environment is appropriately set up to unlock an encrypted Stratis
device by checking if the device can be unlocked with the current
available keys). Checking both at the same time requires the key to
be in the kernel keyring for longer than necessary (for example
during destruction of a pool). This commit reduces the number of
operations that the key is required for to only those that perform
operations that require the key for encryption or decryption.