Preliminary prototype of Report interface
This commit adds a Report interface that can be used to query JSON data
representing various internal engine state information. The JSON schema
is subject to change without being reflected in the semantic version of
stratisd. Only the D-Bus API invocations and return type of a well
formed JSON value will be reflected in the version number. As such, the
Report D-Bus interface is considered unstable. Currently, the only
supported report is that of pools that errored during construction of a
pool on stratisd startup. More reports will be added to this interface
at a later time.