Key handling API
This commit provides a D-Bus API and engine infrastructure to allow
stratisd to manage keys in the kernel keyring securely. This will rule
out the need to send users to keyctl for key management and they will be
able to do so through the stratis CLI or the D-Bus API without leaving
copies of the key in memory locations along the way.

Co-Authored-By: mulhern <amulhern@redhat.com>