Make get_metadata more strict
Its purpose is to be called on a set of already identified device nodes
so return an error if the information on the BDA does not correspond with
what is already known.

It was originally written to be more general, but the purpose of that
generality is long since gone.

Signed-off-by: mulhern <amulhern@redhat.com>