Move handling of liminal devices into its own module
It must get more complex in order to manage the greater complexity of
setup with encryption, so give it the space it needs.

Also, change devlinks methods that explicitly take a Pool as an argument
or return a reference to a Pool, just take a StratPool or return a reference
to a StratPool. The sim engine does not manage devlinks, so it is unnecessary
to make these methods so general.

Ultimately, the devlinks module shouldn't have any reliance on the Pool,
Filesystem, etc. traits, since it is really internal to the strat engine.
I've filed stratisd issue #1987 for that.

Add a FIXME about a possible error condition that we don't attempt to handle.

Signed-off-by: mulhern <amulhern@redhat.com>