Swap ids for Stratis tokens
This is mostly for aesthetic reasons, but it makes sense that the primary
token for a Stratis device be the Stratis token. In future, Stratis may
add management of other tokens for other management approaches, and then it
will seem wierd for the key-descriptor token to be the first token by id.

Signed-off-by: mulhern <amulhern@redhat.com>