Replace most of devlinks module code with udev rule
This commit moves symlinks from /stratis to /dev/stratis by removing
some of the manual devlinks module code in favor of allowing udev to
manage the symlinks that stratisd provides for the filesystems. There
are a few other changes such as moving the mount point for the MDV to
/run/stratisd. This commit should, at a high level, provide tighter
integration with existing Linux infrastructure by adhering more closely
to the filesystem hierarchy and offloading some functionality to udev
instead of implementing it again in stratisd.