Stop defining Debug explicitly
There were two reasons for these explicit definitions:
* The Randomizer in the SimEngine depended on ThreadRng which did not
derive Debug at the time the implementation was needed. But ThreadRng
now defines Debug, and has for a few years, it seems, so this
implementation can be dropped.
* All others were in order to get a simple-ref representation of UUIDs.
I believe that the motivation for this must have been that the engine state
was dumped to the logs simply by using the Debug representation of an
engine. But, we do not dump the engine state like that anymore, since
PR #1907. Furthermore, currently the Debug representations do not have
the intended effect, and instead display the hyphenated representation
anyway. (This may have been due to some change in the underlying uuid crate,
or may been introduced by us in commmit:
fea00efe9596d49f723923ce53924c9e4614974f). It seems that nobody is
complaining about the UUID format in the Debug representations, so I'm
eager to get rid of this non-working extra code, rather than fix it so
that it goes back to working as it was intended to.

I've also added a Display implementation for StratisIdentifiers.

Signed-off-by: mulhern <amulhern@redhat.com>