Remove all expected failure tests
Also, remove all additional support in the tests that they invoke for
encrypted devices.

We will, of course, be adding in lots of new tests for discovery of
encrypted devices. But at present, tests that are expected to fail will
go bad in all sorts of ways without careful monitoring, which is a waste
of time right now.

Signed-off-by: mulhern <amulhern@redhat.com>