Move KeyDescription use up in StratEngine
Move the use of KeyDescription::from() up from initialize() method in crypt
to StratEngine method. This gets the construction of the KeyDescription
directly from a String much closer to the source of the String, and thus
more obviously correct.

It's necessary to make KeyDescription Clone, but the previously used
Strings also had to be cloned to disperse them among the blockdevs and
blockdevmgrs. So there is no additional computational cost.

The implementation is a bit more consistent than previously, as the
KeyDescriptions are passed as references everywhere, until they need to
be cloned to be assigned to their containing data structures.

Becaue it is made explicit, it is more obvious which variant of the key
description is put on the D-Bus.

Signed-off-by: mulhern <amulhern@redhat.com>