Fix several bugs
Do not resize if a good value was found.
Resizing may move the internal buffer, even if the capacity is enough.

Also, calculate the number of keys from the number of bytes read,
which is what the syscall returns.

Also, for reading the description, use as buffer length the capacity, not
the length, which is 0.

Also, dereference the key id, or the pointer, rather than the value, gets
passed.

Signed-off-by: mulhern <amulhern@redhat.com>