Improve some logging for device information manipulation
This means we log for every event that causes any sort of update to
the internal data structures, even if it is a null update, where work
is done but no change actually occurs.

It might make sense to reduce frequency of logging, and make sure only
to log when a real change happens.

Signed-off-by: mulhern <amulhern@redhat.com>