Remove dbus_interface tests
We often forget to add tests, because there is no coverage information to
keep us on top of anything and the individual tests generally end up pretty
perfunctory.

As we improve our other testing, these tests become less and less useful.

So long as they exist they will be a bit of a burden to maintain.

It really does seem like a good time to get rid of them.

Signed-off-by: mulhern <amulhern@redhat.com>