In error message, just print out what was read from the PID file
This is less work than parsing, and substituting "<unknown>" if the
contents of the pid file do not parse properly, and at the same time
a bit more useful. An error will already have been returned if the data is
not valid UTF-8, so it should be printable.

Also, don't write a newline after the PID, there's nothing we want to
put on the next line, so there really isn't any purpose to the newline.

Signed-off-by: mulhern <amulhern@redhat.com>