Drop the hexdump part of stratis_id.py
It dumps a lot of stuff, formatted in an odd way, unconditionally.

If we ever really need to debug Stratis static header regions, we would
instead use tools like hexdump.

Leave in the part that parses the Stratis header, as we have no other
way to get at that at this time via a simple script.

Signed-off-by: mulhern <amulhern@redhat.com>