Have total_physical_used use the cached value of the thinpool status
This has the benefit that the value that stratisd reports is also the
value that it believes. Previously, stratisd could report values that it
did not know about, in the sense that it would not take action about them.
So, stratisd could report that a pool was failed, when its own data
structures indicated otherwise, which would be confusing to a user.

It also has the benefit that when getting this value, stratisd does not
have to run any external command, which makes gathering this information
which may be requested over the D-Bus, more efficient.

Signed-off-by: mulhern <amulhern@redhat.com>