Make a new set_state method and log all state changes there
If we log only on a state change then the log is more useful in two ways.

First, the log is not inundated with repetitions of the same message which
just indicates the state, rather than the change.

Second, if the state changes back to a good state, that change is logged
whereas previously it was not.

Signed-off-by: mulhern <amulhern@redhat.com>