Do not bother to check for extend state in test
It's not useful, because if the pool goes to a not-working state, no
attempt is made to extend the pool, because there is assumed to be not
enough information to determine how to extend the pool. Because of this
there is a fair chance that the pool will never go into any extend
state other than good, as has been oberved to happen in this test.

Signed-off-by: mulhern <amulhern@redhat.com>