Emit D-Bus InterfacesAdded and InterfacesRemoved signals
Currently, stratisd does not emit the appropriate signals when D-Bus
objects are added and removed. This commit provides the necessary
refactoring to support populating the signals with the necessary
information. It also adds support for sending the actions through the
DeferredAction structure that we use for queuing D-Bus actions as
they come in.