Check Info structs for duplicate key descriptions
Remove the check that occurs on the blockdevs after the devices have been
opened. This check is mostly pointless, because if there are multiple key
descriptions, the devices could not have been opened to begin with under
normal operation.

Add a key_description() support method to LInfo to make this more direct.

Signed-off-by: mulhern <amulhern@redhat.com>