Switch to Type=dbus
Type=dbus allows us to specify a name to block on until the stratisd
D-Bus service comes up. systemd will not mark stratisd as up until it
detects this name on the D-Bus. This also implicitly specifies
After=dbus.socket so we are not losing any requirements, only making the
service file more specific.