Deny all the pedantic clippy lints that we already obey
Allow the ones that we don't obey. We obey far more than we disobey.

In due course, remove the allows and change the code for pedantic lints
that we should obey. Comment the others to explain why we don't obey them.

If, through experience, we learn that there is some pedantic lint that we
were obeying, but that is not helpful, add the lint to the list of allowed
pedantic lints, accompanied with a comment explaining why it is allowed.

Signed-off-by: mulhern <amulhern@redhat.com>