Clevis support in stratisd
This commit adds support for NBDE and TPM-based access of encrypted
devices using clevis to bind, unbind, and unlock encrypted devices. It
adds a parameter to the UnlockPool method to specify whether the keyring
or clevis should be used to unlock a pool. It also adds two new methods
on pools: Bind and Unbind. These methods allow users to associate
encrypted pools with a Tang server or TPM.

This commit contains a temporary workaround for cryptsetup's interface
as it does not support reading keys from the kernel keyring. The key is
currently read from the keyring into locked memory and exposed as a file
for consumption via cryptsetup. This code will eventually be removed.

Adding block devices to a pool bound to a TPM or Tang server will cause
the new devices to also be bound to Tang or a TPM with the same
parameters that were originally supplied. This means that adding block
devices requires a network connection to the Tang server in the case of
clevis binding to Tang or the presence of a TPM if the pool is bound to
a TPM. Otherwise adding block devices will fail. If the network goes
down and more storage needs to be added to a Tang-bound pool, the pool
can be unbound from a Tang server temporarily until the network is back
up and unlocked with the existing keyring implementation.