Remove engine event listener infrastructure all together
There was only one implementation of it, the D-Bus one, and since that is
removed, the framework can be removed too. Structures that existed only
because of the event handling support that can also be removed:

The connection field is DbusConnectionData is no longer shared, so it
doesn't have to be reference counted.

The MaybeDbusPath type can be removed, because it was only used for the type
of a dbus path in objects in the engine, and those are no longer necessary
without the event handling mechanism.

Remove a lot of mutability requirements that existed only so that the
dbus path field could be set in internal engine structures.

Signed-off-by: mulhern <amulhern@redhat.com>