No longer iteratively try to set up the D-Bus connection
If the D-Bus connection can not be set up, return with an error.

This simple approach is now acceptable since the systemd configuration file
associated with stratisd makes stratisd of type dbus so that it will not
be started until the D-Bus is available. This is due to the fact that
a service of type D-Bus automatically requires dbus.socket.

This simplifies the MaybeDbusSupport implementation, reducing the number
of method to just three, setup(), process(), and register_pool(). The handle
field of the MaybeDbusSupport struct no longer needs to be optional, since
it is set once, or not at all.

The process_poll() method no longer needs to take an explicit timeout;
it can wait indefinitely for the next event, instead of checking every
second to see if the D-Bus is up until the D-Bus is up.

Signed-off-by: mulhern <amulhern@redhat.com>