Implement DeviceSet and DeviceBag structs
Use in liminal module.

Define a few new convenience methods on LInfo.

Change LInfo::update and LInfo::update_on_remove so that they are private.
Remove a test that required these methods to be public.
Change their arguments to be by reference.
Change their right side arguments to be DeviceInfo instead of LInfo. DeviceInfo
represents the information obtained from the system; LInfo is the information
stored in DeviceSet and DeviceBag.

Change methods in setup.rs so that the HashMap data structures that they take
have a reference to the value type.

Signed-off-by: mulhern <amulhern@redhat.com>