Rewrite update_cargo_crates script
Use MultiDiGraph. This allows tracking the different dependencies that
can arise through different crates.

Print out different results. Use a named tuple for the result type.

Don't try to process the any, all, and not configuration notes.
Better to use some parser for that, probably pyparsing. Not parsing
these appears to have led to one false positive so far, and that was for
a condidtional dependency on libc. It didn't affect the results by
causing libc to be removed because so many others of our libraries depend
on libc.

Compile the regular expressions when the module is loaded, for speed.
Of course, usually pulling down the doc from the koji URL takes almost
all the time, but I guess it's good practice.

Use semantic_version to compare versions.

Don't use the cfg information from rustc to determine relevance of
an entry, just assume that the platform is unix and the target-os is
linux. This is a decision that could be changed in the future, if running
the script on specific configurations. In that case, the URL for the
koji data would also have to be modified.

Add one sanity check: make sure that the disjoint buckets have the same
number of items as the graph has nodes.

Add lots of options to control output.

Require rust- prefix otherwise there can be key collisions in koji repo.

Check against all versions in tree and exit if any exceed Fedora values.

Don't have short form options, there are too many for the short forms to
be mnemonic and argparse is happy with the shortest disambiguating prefix of
any long form anyway.

Signed-off-by: mulhern <amulhern@redhat.com>