Workaround for D-Bus connection hang up
This commit addresses a few issues:

Firstly, this removes the stdin TTY handling on the stratisd side. This
turned out to be too limiting in terms of the scope of file descriptors
that could be passed across the D-Bus to provide a passphrase. Now, any
readable file descriptor may be passed.

Secondly, it is no longer recommended that the user pass stdin as the
file descriptor for interactive activities via the API. Instead, reading
from stdin and passing the result of that read into a pipe or a
pseudoterminal should be strongly considered as the read call on stdin
will block indefinitely if no input is provided and stratisd will
continue to block with no way of terminating the read call. Because
pipes and pseudoterminals have two ends, the end retained by the client
can be closed and the read will terminate on the stratisd side.
Furthermore, passing stdin from a console appears to be blocked by
SELinux. SELinux currently has a rule against providing a daemon with
the file descriptor of a TTY.

Thirdly, the test for whether the passphrase that was provided is too
long is now explicitly non-blocking. The previous method of checking
would block indefinitely if the length of the password provided was
exactly the maximum length of passphrases accepted by stratisd.

One notable change in this commit is the change to the default read
behavior of whatever file descriptor is passed to the SetKey method.
This means that there will be no more custom handling of newlines, etc.
If the read call on the file descriptor returns a newline, that will be
included in the passphrase set in the kernel keyring. If the read
on the file descriptor would block by default and this is not the
desired behavior, it is up to the API client to change the file
descriptor settings to make it non-blocking.