Refactor in preparation for flock PR
This commit makes two major changes to the existing code in preparation
for using flock to guard against storage corruption once multithreading
is implemented.

1. It moves the CryptHandle structure into the StratBlockDev structure.
This is useful for flock as the CryptHandle will now already be acquired
when flock is called allowing the flock code to simply utilize existing
data structures to find the paths of all of the devices to flock as
opposed to having to acquire a CryptHandle each time flock is called on
an encrypted device.

2. It extends BlockDevPath to be able to represent a graph data
structure that models all of our devicemapper devices according to
parent/child relationships. This allows stratisd to traverse all
children down to the leaves of the tree and flock the device in question
as well as all of the children. This will provide additional safety as
we do not want to mutate a devicemapper device and its child at the same
time as this could cause unintended side effects.