Formalize range allocation with a new data structure
The data structure is called PerDevSegments. It takes over the
task of tracking allocated segments and maintaining its invariants.
RangeAllocator continues to be responsible for the algorithm that handles
a request for additional space.

Removes two TODOs by making insertion of a sequence of ranges more
efficient and atomic.

Since PerDevSegments is more of an abstract data structure and can be used
thus, use it as the return type of some methods, instead of using a simple
Vec. This change goes as far as StratBlockDev::request_space(), and thus
affects the one call site in BlockDevMgr, which is adapted.

Adapts all existing tests and add several new tests. States an invariant
on PerDevSegments and uses it in tests.

Adds some additional comments.

Signed-off-by: mulhern <amulhern@redhat.com>