Makefile: Add an install rule
Drain the install logic from the spec file:
https://src.fedoraproject.org/rpms/stratisd/blob/master/f/stratisd.spec#_74
to upstream.

This is better because it keeps the logic for where files go
aligned with the code, not with spec files which live in
a separate git repository on a separate lifecycle.  It allows sharing
installation logic across distributions (and even just Red Hat
has 3 distributions).

It will allow us to avoid the hacky way we `mv` the binaries
after the RPM `%cargo_install` macro.

It allows more easily building from source upstream without
going through RPM at all.

It makes it easier to move towards auto-generated spec
files.

In short: it's better ;)