Make pedantic lints opt-in instead of opt-out
Deny all the existing lints that we already obey.
Explicitly allow all the lints that we think might be helpful.

Add an additional CLIPPY_PEDANTIC_USELESS variable for lints that we have
decided are unhelpful.

We can migrate lints to CLIPPY_PEDANTIC_USELESS if we find that they don't
help.

Rename CLIPPY_DENY to CLIPPY_PEDANTIC.

With this scheme we will not automatically opt in to new lints with each
release.

We will have to drop deprecated lints, but that should be easy since all
lints that we know of are itemized.

We will add new lints to the list only when they seem of great interest
or we have spare time.

We can turn on lints selectively in the CLIPPY_PEDANTIC variable should we
want to explore this.

Signed-off-by: mulhern <amulhern@redhat.com>