Send only one message on name change
Handle all the necessary signals on the DbusTreeHandler side.

Want to minimize the number of signals sent on a finite channel, this
is easy to do since the tree is available to the tree updater.

This also puts the filesystem name change and devnode invalidated
property in the same signal.

Signed-off-by: mulhern <amulhern@redhat.com>