Add ability to create a pool using a Clevis-managed passphrase
Previously, a key in the kernel keyring was always required as a back-up
method for unlocking even when Clevis support was enabled. This commit
adds the ability to create a pool without a key in the kernel keyring,
removes the requirement of a keyring-based passphrase for encrypted
devices being newly initialized, and adds the ability to bind and unbind
from a passphrase in the kernel keyring.