Fix bug in devicemapper event handling
This commit has two parts: fix the bug in the tokio implementation of
devicemapper event polling and resolve the issue where filesystem
creation in a pool does not cause an update of the usage statistics
displayed for the pool.

The bug in the tokio implementation is due to the lock future being
dropped while polling so this commit transitions to an async method
implemented on AsyncFd.

This commit also moves to a data structure that stores the highest event
number for all of the devices in the pool and the number of devicemapper
devices in the pool while also now paying attention to thin device
events. If either the highest event number or number of devicemapper
devices recorded for the pool changes, the thin pool status is
refreshed.