Do not run GitHub Actions tests for certain files
No tests that check anything about README.md or CHANGES.txt exist.

It feels pretty goofy to do all this heavyweight compiling and testing
every week when I update the changelog.

Also, there may come a time when testing is expensive enough that we need
to be more selective about what tests are run for what changes.
For example, we could choose not to run Rust unit tests if all we've done
is updated some Python test code somewhere.
This is a trial run, to see how that goes.

Signed-off-by: mulhern <amulhern@redhat.com>