Fix for hanging synchronization lock in D-Bus layer
This commit resolves an issue where a blocking synchronization lock is
invoked in a few of our methods marked as async by a dependency of stratisd.
This does not always cause a deadlock, but when the tokio scheduler is
heavily used, this can result in commands hanging for a period of time
and resuming or the command timing out all together. The library methods
in question are now always invoked in blocking threads to avoid this
behavior.

Due to the change to blocking threads and method calls in our D-Bus
layer, the should_exit boolean flag has been changed to a channel that
allows us to block on both a message on the channel and the blocking
future for lock acquisition, etc. Receiving a message notifying the
blocking thread to exit will cancel the other future and vice-versa
allowing graceful shutdown of blocking threads as well as non-blocking
threads.

This commit also provides additional logging at the trace level for
synchronization lock acquisition by adding our own locking data
structure that wraps mutex and read-write locks that we use in stratisd.

The last notable change is a custom implementation of GetManagedObjects
that optimizes locking so that the engine mutex does not have to be
acquired when fetching each property. This should remove some
unnecessary overhead in GetManagedObjects.