Add support for changing passphrases and regenerating Clevis bindings
This commit adds two new calls. The first is changing the keying
passphrase for a LUKS2 device. Currently, we support kernel keyring
supplied passphrases, but until now these could not be changed. The new
call will allow changing the passphrase to unlock the LUKS2 devices in
the pool. The second call will trigger a Clevis binding regeneration.
The same Clevis config is passed in, the old passphrase is decrypted, a
new passphrase is set and encrypted, and the old passphrase is removed.

If a passphrase is compromised, these methods can be used to set a new
passphrase for the devices. These calls can also be used for
preemptively rotating passphrases.