No longer use a lowest supported platform linter
Run tests on Fedora 30 instead.

This is not quite our lowest supported environment, just our best proxy.
We can not go lower, because lower Fedora releases do not support the
version of libcryptsetup that stratisd requires.

Since supporting the Python code of these tests so that they can be
run on our lowest supported environment is not necessary, just for
convenience, any extra work on this CI would be wasted.

Signed-off-by: mulhern <amulhern@redhat.com>