Allow disabling actions when stratisd detects unresolvable failures
Previously when failures occurred that caused inconsistencies in the
metadata or other problems that cannot be resolved automatically,
stratisd would not notify the user nor would it prevent any further
actions that could cause the pool to get into an even worse state.

This commit provides two pieces of functionality. Firstly, it adds the
concept of "available actions" for a pool. If a pool attempts to take an
action that is not available given the current state, the pool will
reject it with an error. Secondly, this commit provides the ability to
set the available actions when specific errors are detected that
stratisd knows will leave the metadata or other state in an unworkable
condition.

This will improve the user experience and safety of stratisd moving
forward in error cases, prevent unresolvable issues from popping up, and
allows developers to count on certain parts of the code base starting
with consistent state if they are reached, simplifying the reasoning
needed when looking at certain parts of the code base.