Pass optional size parameter for new filesystem
Change the CreateFilesystems D-Bus call to accept an additional size
parameter.

Export more complete filesystem information in JSON report,
include size and used.

Require filesystem size to be at least 32 MiB and no more than 8 EiB.

Extend SetCreateAction for create filesystems.
Have it include the size of the filesystem, to enrich the log message.

Use an idempotency check that requires that the size and the name of the
filesystem be the same. If the filesystem already exists, but the size is
different, return an error.

Signed-off-by: mulhern <amulhern@redhat.com>