Optimized lock for per-pool locking
This commit provides a locking data strucure that operates at two
different granularities. One is at the level of all pools. The other is
at the level of a single pool. The lock operates according to read-write
lock rules. The result of this will be more aggressive parallelization
for operations that occur on different pools. This commit also migrates
the top level Engine API to asynchronous methods to better enable use in
async methods across IPC layers.