Cache thin dev status when performing check
Use the cached value when determining result of used().

Make the cached value optional, and set its value when initializing
or setting up a filesystem, or when making a snapshot.

Do not require succesfully getting the status when making the snapshot
or setting up a filesystem. The theory is that in many circumstances
users would prefer a filesystem to be set up even if they did not know
the exact usage.

Signed-off-by: mulhern <amulhern@redhat.com>