Transition from greedy to lazy allocation
This commit adds a few notable changes. One is a transition to lazy
allocation with migration code to ensure that older versions of stratisd
still function properly after migration. Another is a filesystem limit
to allow precise calculation of the necessary metadata space for the
thin pool. This ensures that a user can never run out of metadata space
as the amount required is calculated and allocated upfront.