Change liminal device code to respect udev remove events
This commit resolves two issues.

The first is the unnecessary error message on encrypted device destroy.
The message was a warn-level message that could be confusing to users
given that this would be printed every time an encrypted pool was
destroyed and did not indicate that anything had gone wrong.

The second is liminal device code not respecting udev remove events.
The first issue was a symptom of this problem. The code attempts to
check Stratis identifiers on remove events as well as on change and add
events. This results in a failure to read the device in all cases except
for synthetic udev removal events. The event is then ignored and the
internal liminal device code is not updated. The resolution is to rely
on information from the remove udev event to get pool/device UUIDs
instead of reading them from the device.