Clean up D-Bus property diffing infrastructure
This commit primarily enforces the separation between the Stratis pool
data structure and the thin pool implementation, also taking special
care to clearly define used and total space in a way that is helpful to
the user.

Currently, used space is defined as all allocated metadata space, both
Stratis superblock metadata and space taken up by the thin pool metadata
device as well as the number of data device blocks marked as in use.
Total space is the sum of the size of all devices. We have defined it
this way because users should never run out of metadata so we can
consider that all used. Counting data usage and metadata allocation
means that when the data device is full, used space will be equal to
total space. This is desirable from a usability perspective.