Add mode for enabling and disabling overprovisioning
This commit allows users to enable or disable overprovisioning in the
thin provisioning layer for added safety when running out of physical
space should not be an option.

This commit enforces that the total logical size of all filesystems and
snapshots cannot be larger than the amount of physical space available
in the thin pool. Initially, the checks reserve the maximum amount of
metadata space that could possibly be needed for a pool of this size,
but as the data device grows, if this turns out to be an overestimation,
the size of the data device will be used as the amount of space
available to filesystems.

Overprovisioning cannot be disabled if the pool is already
overprovisioned.